# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.CostBreakdown:
    properties:
      base_price_per_hour:
        type: integer
      billed_minutes:
        type: integer
      breakdown_items:
        items:
          $ref: '#/definitions/service.CostItem'
        type: array
      duration_minutes:
        type: integer
      final_cost:
        type: integer
      grace_minutes:
        type: integer
      machine_group_name:
        type: string
      tier_discount_percent:
        type: number
    type: object
  service.CostItem:
    properties:
      amount:
        type: integer
      label:
        type: string
    type: object
...
  service.CreatePrinterRequest:
    properties:
      ip_address:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      port:
        type: integer
      printer_type:
        type: string
      store_id:
        type: string
    required:
    - name
    - printer_type
    type: object
...
  service.EndSessionResponse:
    properties:
      balance_after:
        type: integer
      balance_before:
        type: integer
      cost_breakdown:
        $ref: '#/definitions/service.CostBreakdown'
      duration_minutes:
        type: integer
      machine_code:
        type: string
      machine_id:
        type: string
      member_id:
        type: string
      session_id:
        type: string
      total_cost:
        type: integer
    type: object
...
  service.PrinterResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      port:
        type: integer
      printer_type:
        type: string
      store_id:
        type: string
    type: object
...
  service.SessionDetail:
    properties:
      combo_id:
        type: string
      combo_type:
        type: string
      created_at:
        type: string
      duration_minutes:
        type: integer
      ended_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_overnight:
        type: boolean
      machine_code:
        type: string
      machine_id:
        type: string
      member_id:
        type: string
      member_name:
        type: string
      remaining_minutes:
        type: integer
      slot_end:
        type: string
      started_at:
        type: string
      store_id:
        type: string
      total_cost:
        type: integer
    type: object
...
  service.StartRequest:
    properties:
      combo_purchase_id:
        type: string
      machine_id:
        type: string
      member_id:
        type: string
    required:
    - machine_id
    - member_id
    type: object
...
  service.SwitchMachineRequest:
    properties:
      new_machine_id:
        type: string
    required:
    - new_machine_id
    type: object
...
  service.UpdatePrinterRequest:
    properties:
      ip_address:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      port:
        type: integer
      printer_type:
        type: string
      store_id:
        type: string
    type: object
...
host: localhost:20800
info:
  contact:
    email: support@vnet.net
    name: VNET Support
  description: Hệ thống Quản lý Phòng Game Toàn Diện
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://vnet.net/terms
  title: VNET Core API
  version: "1.0"
paths:
...
  /printers:
    get:
      consumes:
      - application/json
      description: Get a list of all printers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.PrinterResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List all printers
      tags:
      - Printers
    post:
      consumes:
      - application/json
      description: Create a new printer
      parameters:
      - description: Printer data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreatePrinterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.PrinterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new printer
      tags:
      - Printers
  /printers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a printer by ID
      parameters:
      - description: Printer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a printer
      tags:
      - Printers
    get:
      consumes:
      - application/json
      description: Get a single printer by its ID
      parameters:
      - description: Printer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.PrinterResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get printer by ID
      tags:
      - Printers
    put:
      consumes:
      - application/json
      description: Update an existing printer by ID
      parameters:
      - description: Printer ID
        in: path
        name: id
        required: true
        type: string
      - description: Printer update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdatePrinterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.PrinterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a printer
      tags:
      - Printers
  /printers/{id}/test:
    post:
      consumes:
      - application/json
      description: Send a test print job to the printer
      parameters:
      - description: Printer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Test printer connection
      tags:
      - Printers
...
  /sessions/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SessionDetail'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get session by ID
      tags:
      - Sessions
  /sessions/{id}/end:
    post:
      consumes:
      - application/json
      description: End an active play session and calculate charges
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.EndSessionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: End a session
      tags:
      - Sessions
  /sessions/{id}/switch-machine:
    post:
      consumes:
      - application/json
      description: Move an active session to a different machine
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: New machine details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.SwitchMachineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SessionDetail'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Switch machine for a session
      tags:
      - Sessions
  /sessions/active:
    get:
      consumes:
      - application/json
      description: Get all currently active play sessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.SessionDetail'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List active sessions
      tags:
      - Sessions
  /sessions/calculate-cost:
    get:
      consumes:
      - application/json
      parameters:
      - description: Machine ID
        in: query
        name: machine_id
        required: true
        type: string
      - description: Member ID (for discount calculation)
        in: query
        name: member_id
        type: string
      - description: Session duration in minutes
        in: query
        name: duration_minutes
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.CostBreakdown'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Calculate session cost
      tags:
      - Sessions
  /sessions/start:
    post:
      consumes:
      - application/json
      description: Start a new play session for a member on a machine
      parameters:
      - description: Session start details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.StartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SessionDetail'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Start a session
      tags:
      - Sessions
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"